test(cli): promote serve harness + json_report/detail into common (T3.2 prep)
Move ServeClient/send_signal/wait_for/issue_id/id_set out of serve_lifecycle.rs
and migrate_doctor.rs's json_report/detail out of that file, into
tests/common/mod.rs, so shutdown_failure_injection.rs can reuse them without
duplication. Adds an ESRCH-tolerant send_signal_tolerant, a
write_without_reading (background-drains stdout to EOF) on ServeClient, the
shared shutdown-case oracle reopen_and_check, and the pinned bulk_markdown
fixture. No behavioural change to the existing serve_lifecycle.rs /
migrate_doctor.rs test cases.
